Change `LogHistoryContext` to `useLogIncidentHistory` and `useLogReportHistory`
