+++ site/gatsby-site/src/components/incidents/IncidentEditModal.js
@@ -11,7 +11,7 @@ import { useTranslation, Trans } from 'react-i18next';
 import { processEntities } from '../../utils/entities';
 import { getUnixTime } from 'date-fns';
 import { useUserContext } from 'contexts/userContext';
-import { useLogHistoryContext } from 'contexts/LogHistoryContext';
+import { useLogIncidentHistory } from '../../hooks/useLogIncidentHistory';
 
 export default function IncidentEditModal({ show, onClose, incidentId }) {
   const { user } = useUserContext();
@@ -32,7 +32,7 @@ export default function IncidentEditModal({ show, onClose, incidentId }) {
 
   const addToast = useToastContext();
 
-  const { logIncidentHistory } = useLogHistoryContext();
+  const { logIncidentHistory } = useLogIncidentHistory();
 
   useEffect(() => {
     if (incidentData?.incident) {
